fix event listeners
Event listeners now work correctly on completion checkmarks.
Realised that both findIndexByuuid() and findByuuid() are
needed.

DOM creation of checkboxes was refined, instead of innerHTML
proper methods are used

Task class now has edditional conditions on object creation, so
that when tasks marked as completed are retrieved, there is
permanence.